routine_ai/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
PyMySQL>=1.1
pandas>=2.0
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
cryptography>=41
SQLAlchemy>=2.0.0
PyMySQL